Checkpoint

+++ ulib/src/main/java/player/ulib/UNavigation.java
@@ -105,12 +105,15 @@ public class UNavigation {
         return angle;
     }
 
-    //36:24:22 = 36.4061
+    //-------------------------------------------------------------------------
+    // Utility function to convert a DMS string to a Decimal D (float)
+    //Eg. 36:24:22 = 36.4061
+    //
     public static float DMStoD(String dms)
     {
-        int d = Integer.valueOf(dms.split(":")[0]);
-        int m = Integer.valueOf(dms.split(":")[1]);
-        int s = Integer.valueOf(dms.split(":")[2]);
+        float d = Float.valueOf(dms.split(":")[0]);
+        float m = Float.valueOf(dms.split(":")[1]);
+        float s = Float.valueOf(dms.split(":")[2]);
         return (float)d + (float)m/60f + (float)s/3600f;
     }
